Functions for compound distributions.
A compound probability distribution is the probability distribution that results from assuming that a random variable is distributed according to some parametrized distribution, with the parameters of that distribution being assumed to be themselves random variables. Different structures of nested distributions allows to simulate demographic processes. The result of the simulations are probability distributions for different components of a demographic process (e.g. the life long fitness of an individual, the number of reproductive attempts of a cohort).

Currently the binomial (B(n, p)) and negative binomial (N B(k, p)) probability distributions are implemented. The probability parameter p can be a fixed value or a random variable following a Beta distribution with shape parameters α and β (BetaB(n, α, β) and BetaN B(k, α, β)). The size parameter n can be a natural number or a discrete random variable. Finally, k is a positive number or a random variable (not implemented yet).

The implementation represents the distributions by a data frame containing the values and their probability. Thus, it is possible to use recursive compounding and arbitrary distributions. The main handicap arise for discrete distributions with infinite domain (e.g. negative binomial distribution). In this case, one have to omit a part of the probability space to avoid infinite size vectors. By default, negative
binomial distribution is truncated at quantile .99 with a loss of accuracy < 0.01. Accuracy can be increased at expenses of memory and processing time.